requirements.txt
fastapi
uvicorn==0.15.0
python-dotenv==0.19.0
httpx==0.19.0
pydantic
cobo-waas2==1.4.0